+++ src/example/libs/ui/src/upload-toolbar/upload-toolbar.component.ts
@@ -1,0 +1,22 @@
+import { Component, Output, Input, EventEmitter } from "@angular/core";
+import { Upload, Validator } from "@r-hannuschka/ngx-fileupload";
+
+@Component({
+    selector: "app-ui--upload-toolbar",
+    templateUrl: "upload-toolbar.component.html"
+})
+export class UploadToolbarComponent {
+
+    @Input()
+    validator: Validator;
+
+    @Input()
+    url: string;
+
+    @Output()
+    add: EventEmitter<Upload[]> = new EventEmitter();
+
+    public onUploadAdd(uploads: Upload[]) {
+        this.add.emit(uploads);
+    }
+}