+++ src/lib/public-api.ts
@@ -1,12 +1,10 @@
 /*
  * Public API Surface of ngx-fileupload
  */
-export * from "./ngx-fileupload/model/upload";
-export * from "./ngx-fileupload/services/file-upload";
 export * from "./ngx-fileupload/ngx-file-upload.module";
-export * from "./ngx-fileupload/directives/ngx-fileuplad";
-export * from "./ngx-fileupload/components/ngx-fileupload-item.component";
-export * from "./ngx-fileupload/components/ngx-fileupload.component";
-export * from "./ngx-fileupload/validation/validation";
-export * from "./ngx-fileupload/validation/validation.builder";
-export * from "./ngx-fileupload/validation/grouped.validator";
+export * from "@lib/data/api";
+export * from "@lib/data/upload.model";
+export * from "@lib/ui";
+export * from "@lib/utils/http/file-upload";
+export * from "@lib/utils/validation";
+export * from "@lib/utils/upload-control";